% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brundle-data.R
\docType{data}
\name{dbaExperiment}
\alias{dbaExperiment}
\title{Example DiffBind Object}
\format{Diffbind Object}
\source{
Original data from experiment
}
\usage{
jdbaExperiment
}
\description{
This data set is a DiffBind object of example ChIP-seq experiments
}
\author{
Andrew Holding
}
