% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brundle.R
\name{jg.applyNormalisation}
\alias{jg.applyNormalisation}
\title{jg.applyNormalisation}
\usage{
jg.applyNormalisation(jg.experimentPeakset, jg.coefficient, jg.correctionFactor,
  jg.treatedNames)
}
\arguments{
\item{jg.experimentPeakset}{is the peakset extracted from the Diffbind object}

\item{jg.coefficient}{is the coefficient calculated by jg.getNormalizationCoefficient}

\item{jg.correctionFactor}{is the correction factor calculated by jg.getCorrectionFactor}

\item{jg.treatedNames}{is the names of the treated samples}
}
\description{
Takes the experimental peakset and applies the calculated coefficient and
correction factor.
}
\examples{
data(jg.experimentPeakset, package="Brundle")
jg.experimentPeaksetNormalised<-jg.applyNormalisation(jg.experimentPeakset,
                                                       1.267618,
                                                       0.6616886,
                                                       c("1b", "2b", "3b"))

}
\keyword{correction}
\keyword{normalisation}
\keyword{peakset}
