% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brundle-data.R
\docType{data}
\name{jg.conditions}
\alias{jg.conditions}
\title{Example Sample Condition Matrix}
\format{list}
\source{
Original data from experiment
}
\usage{
jg.conditions
}
\description{
This data set is a list of example conditions
}
\author{
Andrew Holding
}
