% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brundle-data.R
\docType{data}
\name{jg.controlCountsTreated}
\alias{jg.controlCountsTreated}
\title{Example ChIP-seq Count Matrix}
\format{matrix}
\source{
Original data from experiment
}
\usage{
jg.controlCountsTreated
}
\description{
This data set is a list counts from an Example ChIP-seq experiment
}
\author{
Andrew Holding
}
