% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brundle-data.R
\docType{data}
\name{jg.controlPeaks}
\alias{jg.controlPeaks}
\title{Example ChIP-seq Control Peakset}
\format{a matrix of locations and counts per experimental sample}
\source{
Original data from experiment
}
\usage{
jg.controlPeaks
}
\description{
This data set is a matrix of example peaks for normalisation of ChIP-seq data
}
\author{
Andrew Holding
}
