% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brundle-data.R
\docType{data}
\name{jg.controlPeakset}
\alias{jg.controlPeakset}
\title{Example ChIP-seq Control Peakset}
\format{a matrix of locations and counts per control sample}
\source{
Original data from experiment
}
\usage{
jg.controlPeakset
}
\description{
This data set is a matrix of example control peaks for normalisation of ChIP-seq data
}
\author{
Andrew Holding
}
