% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brundle-data.R
\docType{data}
\name{jg.controlPeaksetDeSeq}
\alias{jg.controlPeaksetDeSeq}
\title{Example ChIP-seq Control Peakset}
\format{a matrix of locations and counts  sample}
\source{
Original data from experiment
}
\usage{
jg.controlPeaksetDeSeq
}
\description{
This data set is a matrix of example control peaks for normalisation of ChIP-seq data formatted for DeSEQ2
}
\author{
Andrew Holding
}
