% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brundle-data.R
\docType{data}
\name{jg.controlResultsDeseq}
\alias{jg.controlResultsDeseq}
\title{Example ChIP-seq DeSEQ2 Control results}
\format{DeSEQ2 results object}
\source{
Original data from experiment
}
\usage{
jg.controlResultsDeseq
}
\description{
This data set is a GRanges object of example control results from DeSEQ2
}
\author{
Andrew Holding
}
