% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brundle.R
\name{jg.convertPeakset}
\alias{jg.convertPeakset}
\title{jg.convertPeakset}
\usage{
jg.convertPeakset(jg.controlPeakset)
}
\arguments{
\item{jg.controlPeakset}{is the name of the DiffBind object to convert}
}
\description{
Converts a DiffBind object into a DESeq2 compatible form for the workflow.
}
\examples{
jg.convertPeakset(jg.controlPeakset)
}
\keyword{Convert}
\keyword{DESeq2}
\keyword{DiffBind}
