% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brundle.R
\name{jg.correctDBASizeFactors}
\alias{jg.correctDBASizeFactors}
\title{jg.correctDBASizeFactors}
\usage{
jg.correctDBASizeFactors(dba, jg.controlSizeFactors)
}
\description{
Correct the size factors in a DiffBind object using our DESeq2 pipeline for
normalisation.
}
\examples{
jg.correctDBASizeFactors(dba,jg.controlSizeFactors)
}
\keyword{DESeq2}
\keyword{Diffbind}
