% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brundle.R
\name{jg.countAlignedMReads}
\alias{jg.countAlignedMReads}
\title{jg.countAlignedMReads}
\usage{
jg.countAlignedMReads(jg.bamFiles)
}
\arguments{
\item{jg.bamFiles}{is a list of bam files to count.}
}
\description{
This function counts the number of aligned reads in millions from a list of
bam files. It returns these in as a list of numbers in the same order.
}
\keyword{bam}
\keyword{bamFiles}
\keyword{reads}
