% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brundle-data.R
\docType{data}
\name{jg.experimentPeakset}
\alias{jg.experimentPeakset}
\title{Example ChIP-seq Experiment Peakset}
\format{a matrix of locations and counts per control sample}
\source{
Original data from experiment
}
\usage{
jg.experimentPeakset
}
\description{
This data set is a matrix of example peaks for normalisation of ChIP-seq data
}
\author{
Andrew Holding
}
