% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brundle-data.R
\docType{data}
\name{jg.experimentResultsDeseq}
\alias{jg.experimentResultsDeseq}
\title{Example ChIP-seq DeSEQ2 results}
\format{DeSEQ2 results object}
\source{
Original data from experiment
}
\usage{
jg.experimentResultsDeseq
}
\description{
This data set is a GRanges object of example results from DeSEQ2
}
\author{
Andrew Holding
}
