% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brundle.R
\name{jg.getControlCounts}
\alias{jg.getControlCounts}
\title{jg.getControlCounts}
\usage{
jg.getControlCounts(jg.control, jg.controlSampleSheet, jg.Condition)
}
\arguments{
\item{jg.control}{is a peakset extracted by jg.dbaGetPeakset.}

\item{jg.controlSampleSheet}{is the samplesheet supplied to DiffBind to
generate jg.control.}

\item{jg.Condition}{is the condition we the counts for as specficied in the
samplesheet.}
}
\description{
This function counts the number of aligned reads in millions from a list of
bam files. It returns these in as a list of numbers in the same order.
}
\examples{
data(jg.controlPeakset, package="Brundle")
fpath <- system.file("extdata", "samplesheet_SLX14438_hs_CTCF_DBA.csv",package="Brundle")
jg.controlSampleSheet<-fpath
jg.controlCountsTreated<-jg.getControlCounts(jg.controlPeakset, jg.controlSampleSheet,"Fulvestrant")
}
\keyword{bam}
\keyword{bamFiles}
\keyword{reads}
