% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brundle.R
\name{jg.getDba}
\alias{jg.getDba}
\title{jg.getDba}
\usage{
jg.getDba(jg.experimentSampleSheet, dbaSummits, ...)
}
\arguments{
\item{jg.experimentSampleSheet}{is the filename of samplesheet to be loaded}

\item{dbaSummits}{is the peak width in bp from summits (optional)}

\item{...}{are the parameters to be passed to DiffBinds dba.count function}
}
\description{
Generates a DiffBind object from a valid SampleSheet with the required
data for normalisation. No examples are provided as BAM files are not
included in this package.
}
\keyword{DiffBind}
\keyword{counts}
\keyword{load}
\keyword{samplesheet}
