% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brundle.R
\name{jg.getNormalizationCoefficient}
\alias{jg.getNormalizationCoefficient}
\title{jg.getNormalizationCoefficient}
\usage{
jg.getNormalizationCoefficient(jg.controlCountsTreated,
  jg.controlCountsUntreated)
}
\arguments{
\item{jg.controlCountsTreated}{Control counts extracted from the Diffbind object for the treated condition using jg.getControlCounts}

\item{jg.controlCountsUntreated}{Control ounts extracted from the Diffbind object for the untreated condition using jg.getControlCounts}
}
\description{
This function allows the user to caryy out the normalisation and returns a
coefficient by using a linear fit to the control data.
}
\examples{
data(jg.controlCountsTreated, package="Brundle")
data(jg.controlCountsUntreated, package="Brundle")
jg.coefficient<-jg.getNormalizationCoefficient(jg.controlCountsTreated,
                                                  jg.controlCountsUntreated)
}
\keyword{normalization}
