% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brundle.R
\name{jg.getSampleIds}
\alias{jg.getSampleIds}
\title{jg.getSampleIds}
\usage{
jg.getSampleIds(jg.controlSampleSheet)
}
\arguments{
\item{jg.controlSampleSheet}{is the filename of the samplesheet}
}
\description{
Extracts the sample Id from DiffBind formatted SampleSheet in csv format.
}
\keyword{DiffBind}
\keyword{sample}
\keyword{samplesheet}
