% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brundle.R
\name{jg.plotDeSeq}
\alias{jg.plotDeSeq}
\title{jg.plotDeSeq}
\usage{
jg.plotDeSeq(ma.df, p = 0.01, title.main = "Differential ChIP",
  log2fold = 0.5, flip = FALSE)
}
\arguments{
\item{ma.df}{is the result Dataframe from DESeq2}

\item{p}{is the minium FDR to highlight as significant}

\item{log2fold}{is the minimum log2 fold change for highlighted points}

\item{flip}{when set to TRUE flips the data}

\item{title}{is the plot title}
}
\description{
Plots the output from DESeq2 for the Brundle pipeline
}
\examples{
 data(jg.experimentResultsDeseq,package="Brundle")
 jg.plotDeSeq(jg.experimentResultsDeseq,
  p=0.01,
  title.main="Fold-change in ER binding",
  flip=TRUE
 )

}
\keyword{DESeq2}
\keyword{data}
\keyword{plot}
