% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brundle.R
\name{jg.plotMA}
\alias{jg.plotMA}
\title{jg.plotMA}
\usage{
jg.plotMA(jg.experimentPeakset, jg.controlPeakset, jg.untreatedNames,
  jg.treatedNames, jg.coefficient)
}
\arguments{
\item{jg.experimentPeakset}{is the peakset of the experimental data extracted from a DiffBind ojbect with jg.dbaGetPeakset}

\item{jg.controlPeakset}{is the peakset of the control data extracted from a DiffBind ojbect with jg.dbaGetPeakset}

\item{jg.untreatedNames}{is a list of sample names for the control or untreated conditions}

\item{jg.treatedNames}{is a list of sample samples for the treated conditions}

\item{jg.coefficient}{is a normalisation coefficent for the data that can be generated via the pipeline. Can be set to 1 to view before normalisation.}
}
\description{
This function plots both the control and experimental data on an MA plot.
It also allows for the user to provide a normalisation coefficent for the data.
}
\keyword{normalization}
\keyword{plot}
