% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brundle.R
\name{jg.plotNormalization}
\alias{jg.plotNormalization}
\title{jg.plotNormalization}
\usage{
jg.plotNormalization(jg.controlCountsTreated, jg.controlCountsUntreated)
}
\arguments{
\item{jg.controlCountsTreated}{Control counts extracted from the Diffbind object for the treated condition using jg.getControlCounts}

\item{jg.controlCountsUntreated}{Control ounts extracted from the Diffbind object for the untreated condition using jg.getControlCounts}
}
\description{
This function allows the user to visualize the normalisation. It is not
needed for the pipeline but provides a helpful illustration of the process.
}
\examples{
data(jg.controlCountsTreated, package="Brundle")
data(jg.controlCountsUntreated, package="Brundle")
jg.plotNormalization(jg.controlCountsTreated, jg.controlCountsUntreated)
}
\keyword{normalization}
\keyword{plot}
