% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brundle.R
\name{jg.runDeSeq}
\alias{jg.runDeSeq}
\title{jg.runDeSeq}
\usage{
jg.runDeSeq(jg.PeaksetDeSeq, jg.conditions, jg.SizeFactors = NULL)
}
\arguments{
\item{jg.PeaksetDeSeq}{is the experimental peakset formatted for DESeq2}

\item{jg.conditions}{is the list of conditions to be compared}

\item{jg.SizeFactors}{is the size factors generated from the control samples}
}
\description{
Runs DESeq2 on our peakset after we have obtained the normalised size factors.
}
\examples{
jg.runDeSeq(jg.PeaksetDeSeq,jg.conditions, jg.SizeFactors = NULL)
}
\keyword{DESeq2}
